# Trustify.Backend/Trustify.Backend.AdminService/Controllers/UserIdGroupWrapper.cs
namespace Trustify.Backend.AdminService.Controllers
{
    public class UserIdGroupWrapper
    {
        public string UserId { get; set; } = null!;

        public string GroupId { get; set; } = null!;
    }
}